add solve for 2447 in C langs
Reviewed by Wizley.
2019.12.31.

I added solve for 2447 prob in C language.

First, I tried to find out what this problem really is line by line but
wast not that helpful.

I used Divide & Conquer algorithm to solve this problem recursively. It
gets input from user and call drawXYwithSize recursively while size is
1 which means we can draw star in that index.

For example of size 27, first It divide 27x27 map to 9(27/3) cubes and
call drawXYwithSize recursively. the new_size becomes 3 and finally 1.
We just need skip the case when i, j of function is 1 in case of middle
loc.