add solve for 1463 in C++ langs
Reviewed by Wizley.
2020.01.03.

I added solve for 1463 in C++ language.

I used DP alogorithm to solve this prob. First I just check the
condition whether it is dividable by 3, 2. But It was not the best
condition. So I used DP algorithm and saved the result of each number
and verfiy the best case for each number.