add solve for 10814 in C++ langs
Reviewed by Wizley.

I have added solve for 10814 in C++ languages.
I stucked in time limit problem and was with "endl".
By using endl the program has to get rid of buffer every time, so it is much faster to use "\n" instead of "endl". and also, to have no problem of sorting same priority, used stable_sort.