add solve for 1406 in C++ langs
Reviewed by Wizley.
2020.01.30.

I have added solve for 1406 in C++ languages. I used list container to solve this prob. and had runtime error 3 times :(. It was caused by erase method of list. By erasing the element, we must care for the cursor's idx by saving return value of erase to that of value.